cleaning pt. 5 - renaming and structuing classes
